fix: Readme and Webpack
