+++ babel.config.js
@@ -2,8 +2,13 @@
 module.exports = {
   presets: [
     ["@babel/preset-env", { targets: { node: "current" } }],
-    "@babel/preset-react",
+    [
+      "@babel/preset-react",
+      {
+        runtime: "automatic",
+      },
+    ],
     "@babel/preset-typescript",
   ],
-  plugins:     ["@babel/proposal-class-properties"]
+  plugins: ["@babel/proposal-class-properties"],
 };
