+++ tsconfig.json
@@ -5,7 +5,7 @@
     "module": "es2020",
     "sourceMap": true,
     "target": "es6",
-    "jsx": "react",
+    "jsx": "react-jsxdev",
     "allowJs": false,
     "moduleResolution": "node",
     "esModuleInterop": true,
